feat(backend): moved tests back to models
